VUE-523 renamed kv-dropdown-rounded to kv-select
