+++ src/pages/Autolending/LoanIncrementDropdown.vue
@@ -3,7 +3,7 @@
 		<h3 class="filter-title">
 			Loan sizes
 		</h3>
-		<kv-dropdown-rounded v-model="loanIncrement">
+		<kv-select v-model="loanIncrement">
 			<option value="any">
 				Any amounts
 			</option>
@@ -28,19 +28,19 @@
 			<option value="1000">
 				$1000
 			</option>
-		</kv-dropdown-rounded>
+		</kv-select>
 	</div>
 </template>
 
 <script>
 import _get from 'lodash/get';
 import gql from 'graphql-tag';
-import KvDropdownRounded from '@/components/Kv/KvDropdownRounded';
+import KvSelect from '@/components/Kv/KvSelect';
 
 export default {
 	inject: ['apollo', 'cookieStore'],
 	components: {
-		KvDropdownRounded,
+		KvSelect,
 	},
 	data() {
 		return {
